+++ CHANGELOG.md
@@ -2,3 +2,7 @@
 
 #### v1.0.0:
 - Add settings and runtime scriptables
+
+#### v1.0.1:
+- Fix cyclic references
+- Fix Scriptable Settings Print